+++ src/pages/index.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Box, Container, Hidden, makeStyles, Typography } from '@material-ui/core'
 
-import ButtonLink from '../components/buttonlink'
+import { ButtonLink } from '../components/links'
 import Image from "../components/image"
 import SEO from "../components/seo"
 
@@ -12,11 +12,14 @@ const useStyles = makeStyles(theme => ({
     flexGrow: 1,
     height: '100vh'
   },
-  links: {
+  linkCol: {
     flexGrow: 1,
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center'
+  },
+  btnLink: {
+    marginBottom: theme.spacing(1)
   },
   image: {
     flexGrow: 3
@@ -42,9 +45,10 @@ const IndexPage = () => {
     <SEO title="Home" />
     <Hidden smDown>
       <Box className={classes.root}>
-        <Box className={classes.links}>
+        <Box className={classes.linkCol}>
           <LinkSectionHeader className={classes.linksectheaderbox} />
-          <Box><ButtonLink variant="outlined" to="/calc">Calculator</ButtonLink></Box>
+          <ButtonLink className={classes.btnLink} variant="outlined" to="/calc">Calculator</ButtonLink>
+          <ButtonLink className={classes.btnLink} variant="outlined" to="/planner">Game Planner</ButtonLink>
         </Box>
         <Box className={classes.image}><Image /></Box>
       </Box>
