+++ src/pages/planner.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import Layout from '../components/layout'
+
+const Planner = () => {
+    return (<Layout>
+
+    </Layout>)
+}
+
+export default Planner
